+++ src/main/java/com/petplate/petplate/petfood/domain/entity/BookMarkedFeed.java
@@ -34,7 +34,7 @@ public class BookMarkedFeed extends BaseEntity {
     private float serving;
 
     @Column(nullable = false)
-    private float kcal;
+    private double kcal;
 
     @Embedded
     private Nutrient nutrient;
@@ -44,7 +44,7 @@ public class BookMarkedFeed extends BaseEntity {
     private User user;
 
     @Builder
-    public BookMarkedFeed(String name, float serving, float kcal, Nutrient nutrient, User user) {
+    public BookMarkedFeed(String name, float serving, double kcal, Nutrient nutrient, User user) {
         this.name = name;
         this.serving = serving;
         this.kcal = kcal;
